Chore: Move onHover handler to its own file for better testing
